Files/FreeLancer Website/VAS WEBSITE/userproject/requirements.txt
altair==5.5.0
altgraph==0.17.4
asgiref==3.8.1
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.1.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
crispy-bootstrap5==0.7
cryptography==45.0.5
cycler==0.12.1
debugpy==1.8.15
decorator==5.2.1
defusedxml==0.7.1
diff-match-patch==20241021
dj-database-url==3.0.1
Django==5.0.2
django-admin-interface==0.30.1
django-allauth==0.60.1
django-colorfield==0.14.0
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-environ==0.11.2
django-import-export==4.3.8
django-widget-tweaks==1.5.0
djangorestframework==3.15.0
executing==2.2.0
fonttools==4.59.0
future==1.0.0
geocoder==1.38.1
gitdb==4.0.12
GitPython==3.1.45
gunicorn==21.2.0
idna==3.10
imbalanced-learn==0.13.0
ipykernel==6.30.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mysql-connector-python==9.4.0
mysqlclient==2.2.7
narwhals==2.0.1
nest-asyncio==1.6.0
numpy==2.2.0
oauthlib==3.3.1
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.1
parso==0.8.4
pefile==2023.2.7
pillow==11.2.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
protobuf==6.31.1
psutil==7.0.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
pyarrow==21.0.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.19.2
pyinstaller==6.14.1
pyinstaller-hooks-contrib==2025.5
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-decouple==3.8
python-slugify==8.0.4
python3-openid==3.2.0
pytz==2025.2
# pywin32==311
# pywin32-ctypes==0.2.3
pyzbar==0.1.9
pyzmq==27.0.0
qrcode==8.2
ratelim==0.1.6
referencing==0.36.2
requests==2.32.4
requests-oauthlib==2.0.0
rpds-py==0.26.0
scikit-learn==1.6.1
scipy==1.16.1
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
sklearn-compat==0.1.3
smmap==5.0.2
sqlparse==0.5.3
stack-data==0.6.3
streamlit==1.47.1
stripe==8.4.0
tablib==3.8.0
tenacity==9.1.2
text-unidecode==1.3
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.1
traitlets==5.14.3
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
wcwidth==0.2.13
wheel==0.45.1
whitenoise==6.6.0
